HGI-8978: Force self.count to be an integer

+++ tap_salesforce/client.py
@@ -242,7 +242,7 @@ class SalesforceStream(RESTStream):
         ):
             msg = self.response_error_message(response)
             count = self.config.get("order_page_size") if hasattr(self.config,"order_page_size") else self.count if hasattr(self,"count") else 200
-            self.count = count / 2
+            self.count = int(count / 2)
             raise RetriableAPIError(msg, response)
         try:
             res_json = response.json()
